data_collector.py
#!/usr/bin/env python3
"""
Egg Price Data Collector
Collects egg price data from various sources and stores it in the database
"""

import os
import requests
import psycopg2
from datetime import datetime, date
import json
import time
import logging
import random
from typing import Dict, List, Optional

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

class EggPriceCollector:
    def __init__(self):
        self.db_config = {
            'host': os.getenv('DB_HOST', 'localhost'),
            'port': os.getenv('DB_PORT', '5432'),
            'user': os.getenv('DB_USER', 'postgres'),
            'password': os.getenv('DB_PASSWORD', 'password'),
            'database': os.getenv('DB_NAME', 'egg_tracker')
        }
        self.conn = None
    
    def connect_db(self):
        """Connect to PostgreSQL database"""
        try:
            self.conn = psycopg2.connect(**self.db_config)
            logger.info("Database connection established")
        except Exception as e:
            logger.error(f"Failed to connect to database: {e}")
            raise
    
    def close_db(self):
        """Close database connection"""
        if self.conn:
            self.conn.close()
            logger.info("Database connection closed")
    
    def insert_price_data(self, price_data: Dict):
        """Insert price data into database"""
        if not self.conn:
            self.connect_db()
        
        try:
            cur = self.conn.cursor()
            query = """
                INSERT INTO egg_prices (date, location, price_per_dozen, source)
                VALUES (%(date)s, %(location)s, %(price_per_dozen)s, %(source)s)
                ON CONFLICT (date, location) DO UPDATE SET
                price_per_dozen = EXCLUDED.price_per_dozen,
                source = EXCLUDED.source
            """
            cur.execute(query, price_data)
            self.conn.commit()
            cur.close()
            logger.info(f"Inserted price data for {price_data['location']} on {price_data['date']}")
        except Exception as e:
            logger.error(f"Failed to insert price data: {e}")
            if self.conn:
                self.conn.rollback()
    
    def insert_location(self, location_name: str, location_type: str):
        """Insert location into database if it doesn't exist"""
        if not self.conn:
            self.connect_db()
        
        try:
            cur = self.conn.cursor()
            query = """
                INSERT INTO locations (name, type)
                VALUES (%s, %s)
                ON CONFLICT (name) DO NOTHING
            """
            cur.execute(query, (location_name, location_type))
            self.conn.commit()
            cur.close()
            logger.info(f"Location {location_name} ensured in database")
        except Exception as e:
            logger.error(f"Failed to insert location: {e}")
            if self.conn:
                self.conn.rollback()
    
    def simulate_usda_data(self) -> float:
        """
        Simulate USDA national average data
        In a real implementation, this would fetch from USDA API or scrape their website
        """
        # Simulate realistic egg price fluctuations around $2.00-$3.00
        base_price = 2.25
        fluctuation = random.uniform(-0.30, 0.50)
        return round(base_price + fluctuation, 2)
    
    def simulate_local_prices(self, locations: List[str]) -> Dict[str, float]:
        """
        Simulate local price data for various locations
        In a real implementation, this would scrape local grocery store websites
        """
        prices = {}
        national_base = 2.25
        
        # Different multipliers for different regions
        location_multipliers = {
            'California': 1.15,      # Higher cost of living
            'Texas': 0.90,           # Lower cost of living
            'New York': 1.25,        # Higher cost of living
            'Florida': 0.95,         # Moderate cost of living
            'Illinois': 1.05,        # Moderate cost of living
        }
        
        for location in locations:
            multiplier = location_multipliers.get(location, 1.0)
            base = national_base * multiplier
            fluctuation = random.uniform(-0.25, 0.35)
            prices[location] = round(base + fluctuation, 2)
        
        return prices
    
    def collect_fred_data(self) -> Optional[float]:
        """
        Collect data from FRED (Federal Reserve Economic Data)
        This is a placeholder - you'd need to implement actual FRED API calls
        """
        try:
            # This would be a real API call in production
            # fred_api_key = os.getenv('FRED_API_KEY')
            # response = requests.get(f'https://api.stlouisfed.org/fred/series/observations?series_id=APU0000708111&api_key={fred_api_key}&file_type=json')
            
            # For now, return simulated data
            return self.simulate_usda_data()
        except Exception as e:
            logger.error(f"Failed to collect FRED data: {e}")
            return None
    
    def collect_grocery_store_data(self, locations: List[str]) -> Dict[str, float]:
        """
        Collect data from grocery store websites
        This is a placeholder for web scraping implementation
        """
        try:
            # In production, this would scrape actual grocery store websites
            # Example stores: Walmart, Kroger, Safeway, etc.
            return self.simulate_local_prices(locations)
        except Exception as e:
            logger.error(f"Failed to collect grocery store data: {e}")
            return {}
    
    def run_collection(self):
        """Run the complete data collection process"""
        logger.info("Starting egg price data collection")
        
        try:
            self.connect_db()
            
            # Collect national average
            national_price = self.collect_fred_data()
            if national_price:
                self.insert_price_data({
                    'date': date.today(),
                    'location': 'NATIONAL',
                    'price_per_dozen': national_price,
                    'source': 'USDA/FRED'
                })
            
            # Collect local prices
            locations = ['California', 'Texas', 'New York', 'Florida', 'Illinois']
            local_prices = self.collect_grocery_store_data(locations)
            
            for location, price in local_prices.items():
                # Ensure location exists in database
                self.insert_location(location, 'state')
                
                # Insert price data
                self.insert_price_data({
                    'date': date.today(),
                    'location': location,
                    'price_per_dozen': price,
                    'source': 'Local Market Data'
                })
            
            logger.info("Data collection completed successfully")
            
        except Exception as e:
            logger.error(f"Data collection failed: {e}")
        finally:
            self.close_db()

def main():
    """Main function for command line execution"""
    collector = EggPriceCollector()
    collector.run_collection()

if __name__ == "__main__":
    main()
    